validation ready

+++ lib/libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: rdel-fra <rdel-fra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/07 15:43:54 by rdel-fra          #+#    #+#             */
-/*   Updated: 2025/03/14 14:56:28 by rdel-fra         ###   ########.fr       */
+/*   Updated: 2025/07/25 15:55:18 by rdel-fra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i] != '\0')
